move BC validation to base validator
