+++ demisto_sdk/commands/common/hook_validations/content_entity_validator.py
@@ -5,6 +5,7 @@ from abc import abstractmethod
 from distutils.version import LooseVersion
 from typing import Optional
 
+import click
 from packaging import version
 
 from demisto_sdk.commands.common.constants import (
@@ -25,6 +26,8 @@ from demisto_sdk.commands.common.tools import (_get_file_id, find_type,
                                                get_file_displayed_name,
                                                is_test_config_match,
                                                run_command)
+from demisto_sdk.commands.format.format_constants import \
+    OLD_FILE_DEFAULT_1_FROMVERSION
 
 json = JSON_Handler()
 yaml = YAML_Handler()
@@ -60,7 +63,21 @@ class ContentEntityValidator(BaseValidator):
             self.are_fromversion_and_toversion_in_correct_format(),
             self.are_fromversion_toversion_synchronized(),
         ]
+
         return all(tests)
+
+    def is_backward_compatible(self):
+        if not self.old_file:
+            return True
+
+        click.secho(f'Validating backwards compatibility for {self.file_path}')
+
+        is_bc_broke = [
+            self.is_id_not_modified(),
+            self.is_valid_fromversion_on_modified(),
+        ]
+
+        return not any(is_bc_broke)
 
     def is_valid_generic_object_file(self):
         tests = [
@@ -72,6 +89,53 @@ class ContentEntityValidator(BaseValidator):
     def is_valid_version(self):
         # type: () -> bool
         pass
+
+    @error_codes('BC105')
+    def is_id_not_modified(self):
+        # type: () -> bool
+        """Check if the ID of the file has been changed.
+
+        Returns:
+            (bool): Whether the file's ID has been modified or not.
+        """
+        if not self.old_file:
+            return True
+
+        old_version_id = self.structure_validator.get_file_id_from_loaded_file_data(self.old_file)
+        new_file_id = self.structure_validator.get_file_id_from_loaded_file_data(self.current_file)
+        if not (new_file_id == old_version_id):
+            error_message, error_code = Errors.file_id_changed(old_version_id, new_file_id)
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                return False
+
+        # True - the id has not changed.
+        return True
+
+    @error_codes('BC106')
+    def is_valid_fromversion_on_modified(self):
+        # type: () -> bool
+        """Check that the fromversion property was not changed on existing Content files.
+
+        Returns:
+            (bool): Whether the files' fromversion as been modified or not.
+        """
+        if not self.old_file:
+            return True
+
+        from_version_new = self.current_file.get("fromversion") or self.current_file.get("fromVersion")
+        from_version_old = self.old_file.get("fromversion") or self.old_file.get("fromVersion")
+
+        # if in old file there was no fromversion ,format command will add from version key with 4.1.0
+        if not from_version_old and from_version_new == OLD_FILE_DEFAULT_1_FROMVERSION:
+            return True
+
+        if from_version_old != from_version_new:
+            error_message, error_code = Errors.from_version_modified()
+            if self.handle_error(error_message, error_code, file_path=self.file_path):
+                self.is_valid = False
+                return False
+
+        return True
 
     @error_codes('BA100')
     def _is_valid_version(self):
